requirements.txt
# Database and environment dependencies
pymysql==1.1.0
python-dotenv==1.0.0

# HTTP requests dependency
requests==2.31.0

# Kafka dependencies
confluent-kafka==2.3.0

# Additional dependencies
mysql-connector-python==8.2.0


# cmd to install all the dependencies:
# pip3 install -r requirements.txt